\name{Bioc2020Anno-package}
\alias{Bioc2020Anno-package}
\alias{Bioc2020Anno}
\docType{package}
\title{
\packageTitle{Bioc2020Anno}
}
\description{
\packageDescription{Bioc2020Anno}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Bioc2020Anno}
\packageIndices{Bioc2020Anno}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Bioc2020Anno}

Maintainer: \packageMaintainer{Bioc2020Anno}
}

\keyword{ package }
